First set of tests passing with new t-cycle-accurate CPU.
Well, LD r, r is passing. Still, pretty huge.
This rewrite (again, sorry) of the main decode/execute block is after me spending weeks working on a Verilog version of the CPU. I learned a lot, and decided to use rust as sort of a verification/code generation system. The CPU is designed in a spreadsheet (ha!). The sheet is imported, then "compiled" into high-level micro-code, which is then "decoded" into the actual micro-code that I'll be using in Verilog.