Working sound on windows after changes to libsoundio.
Apparently libsoundio relies on the system timer to synchronize the callbacks, which defaults to 15ms resolution on Windows. I noticed this after I realized the callback was being called suspiciously at 15ms intervals.
The fix is to set the system timer to 1ms intervals (which isn't enough), and to wait for 2ms less than usual. Still doesn't take us all the way to 32 samples per frame, but oh well.